requirements.txt
pillow>=9.1.1